requirements.txt
pytest
selenium
playwright
pytest-playwright
webdriver_manager
allure-pytest
allure-python-commons